Refactored rake as per new model
